Upgrade to JFlex 1.8.2

+++ opengrok-indexer/src/main/java/org/opengrok/indexer/analysis/DisjointSpanChangedEvent.java
@@ -18,7 +18,7 @@
  */
 
 /*
- * Copyright (c) 2017, Chris Fraire <cfraire@me.com>.
+ * Copyright (c) 2017, 2020, Chris Fraire <cfraire@me.com>.
  */
 
 package org.opengrok.indexer.analysis;
@@ -31,7 +31,7 @@ public class DisjointSpanChangedEvent {
 
     private final Object source;
     private final String className;
-    private final int position;
+    private final long position;
 
     /**
      * Initializes an immutable instance of {@link DisjointSpanChangedEvent}.
@@ -39,8 +39,7 @@ public class DisjointSpanChangedEvent {
      * @param className the span class name
      * @param position the span position
      */
-    public DisjointSpanChangedEvent(Object source, String className,
-        int position) {
+    public DisjointSpanChangedEvent(Object source, String className, long position) {
         this.source = source;
         this.className = className;
         this.position = position;
@@ -66,7 +65,7 @@ public class DisjointSpanChangedEvent {
      * Gets the span position.
      * @return the initial value
      */
-    public int getPosition() {
+    public long getPosition() {
         return position;
     }
 }
